ajout route_aurelia
